delete unused fields
