+++ internal/usecase/interfaces/scene.go
@@ -18,7 +18,6 @@ var (
 type Scene interface {
 	Fetch(context.Context, []id.SceneID, *usecase.Operator) ([]*scene.Scene, error)
 	FindByProject(context.Context, id.ProjectID, *usecase.Operator) (*scene.Scene, error)
-	FetchLock(context.Context, []id.SceneID, *usecase.Operator) ([]scene.LockMode, error)
 	Create(context.Context, id.ProjectID, *usecase.Operator) (*scene.Scene, error)
 	AddWidget(context.Context, id.SceneID, id.PluginID, id.PluginExtensionID, *usecase.Operator) (*scene.Scene, *scene.Widget, error)
 	UpdateWidget(context.Context, UpdateWidgetParam, *usecase.Operator) (*scene.Scene, *scene.Widget, error)
